scheduler: call `schedule()` only when needed
Previously, we were calling the `schedule()` method on every step of the
simulation. This was time consuming and prevented us from simulating
data from large-ish traces.

This set of changes makes sure we only call `schedule` when we actually
need it. Namely, when new jobs are submitted, or when the time of the
next event in the simulation matches the current simulation step.

Processing events might cause the status of the system to change and,
therefore, the scheduling algorithm might need to update its data
structures.