code(env): improve support for trace-based workloads
Using traces was awkward, at best, since we had to provide a time limit
to be used.

This commit fixes this by allowing time_limit to be zero, in which case
the environment will default (on each reset) to use the submission time
of the last job plus its execution time to set as time limit.

This is not ideal, though, as in some high-load cases we will lose many
jobs at the tail of the workload.

Still, this is appropriate for now.