code(env): add yet another reward computation method
After some careful thinking, the job-slots method might
be too greedy, so we're adding another method that computes
rewards based both on jobs in job-slots, and running jobs.

This might give agents the opportunity to observe the impact
of past decisions (running jobs) for a while, without adding
the noise of jobs in the whole waiting queue.