code(env): add support for intermediate rewards
One impact of using the compact MDP is that the MDP automatically
becomes a semi-MDP, with time-varying actions. When we do so, we
might need intermediate rewards to compute the actual reward of
taking an action (for example, the sum of intermediate rewards).

With the size of the sequence, we can compute the true discount factor
as well, if a discount factor is used.